requirements.txt
langchain
openai
pinecone-client
torch
pdf2image